Rename RifColorLegendData to RifFormationNamesReader
